sde_sampler bug fixed

+++ cosmos_predict2/_src/predict2/rl/grpo_sde_sampler.py
@@ -107,8 +107,11 @@ def grpo_sde_step(
     else:
         next_latents = fixed_next_latents.to(torch.float32)
 
-    log_prob = _normal_log_prob(next_latents, prev_sample_mean, std)
+    next_latents = next_latents.to(latents.dtype)
+    # Compute log_prob on the exact value that callers cache/reuse, so rollout old_log_prob
+    # and training-time new_log_prob evaluate the same transition.
+    log_prob = _normal_log_prob(next_latents.to(torch.float32), prev_sample_mean, std)
     # print(f"mean: {mean[0, 0, 0, :10, 10]}")
     # print(f"std: {std}")
     # print(f"log_probs: {log_prob}")
-    return GrpoStepOutput(next_latents=next_latents.to(latents.dtype), pred_x0=pred_x0.to(latents.dtype), log_prob=log_prob)
+    return GrpoStepOutput(next_latents=next_latents, pred_x0=pred_x0.to(latents.dtype), log_prob=log_prob)
